requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
python-decouple>=3.8
django-filter>=23.0
mysqlclient>=2.0  # Para PythonAnywhere MySQL
whitenoise>=6.0  # Para servir arquivos estáticos